# api/swagger-spec.yaml
openapi: 3.0.0
paths:
  /api/shops:
    post:
      operationId: ShopsController_create
      summary: 物販を作成する
      description: 物販を作成します。カレントユーザーを物販のオーナー販売者として設定します。
      parameters:
        - &ref_0
          name: x-user-id
          in: header
          description: 認証モジュールを実装したら代わりにAuthorizationヘッダーにトークンを載せて送ってもらう予定
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShopDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseShopDto'
        '400':
          description: ''
      tags: &ref_1
        - Shop / 物販
    get:
      operationId: ShopsController_findAll
      summary: 物販のリストを取得する
      description: カレントユーザーが販売者として所属する物販のリストを取得します。
      parameters:
        - *ref_0
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseShopListDto'
        '400':
          description: ''
      tags: *ref_1
  /api/shops/{shopId}:
    get:
      operationId: ShopsController_findOne
      summary: 物販を取得する
      description: 物販を1つ取得します。取得できる物販はカレントユーザーが販売者として所属するものに限ります。
      parameters:
        - *ref_0
        - name: shopId
          required: true
          in: path
          example: 1
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseShopDto'
        '400':
          description: ''
      tags: *ref_1
    patch:
      operationId: ShopsController_update
      summary: 物販を更新する
      description: 物販を更新します。更新できる物販はカレントユーザーが販売者として所属するものに限ります。
      parameters:
        - *ref_0
        - name: shopId
          required: true
          in: path
          example: 2
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShopDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseShopDto'
        '400':
          description: ''
      tags: *ref_1
    delete:
      operationId: ShopsController_delete
      summary: 物販を削除する
      description: 物販を削除します。削除できる物販はカレントユーザーが販売者として所属し、かつオーナー権限を持つものに限ります。
      parameters:
        - *ref_0
        - name: shopId
          required: true
          in: path
          example: 1
          schema:
            type: number
      responses:
        '204':
          description: ''
        '400':
          description: ''
      tags: *ref_1
  /api/shops/{shopId}/merchs:
    get:
      operationId: MerchsController_findAll
      summary: マーチのリストを取得する
      description: マーチのリストを取得します。
      parameters:
        - &ref_2
          name: x-user-id
          in: header
          description: 認証モジュールを実装したら代わりにAuthorizationヘッダーにトークンを載せて送ってもらう予定
          schema:
            type: string
        - name: cursor
          required: false
          in: query
          description: ページネーションのカーソル。API は cursor で渡されたidより後のレコードを返します
          example: 34
          schema:
            default: 1
            type: number
        - name: take
          required: false
          in: query
          description: 取得件数
          example: 50
          schema:
            maximum: 1000
            default: 200
            type: number
        - name: sortBy
          required: false
          in: query
          description: ソート
          example: asc
          schema:
            default: asc
            enum:
              - asc
              - desc
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMerchListDto'
        '400':
          description: ''
      tags: &ref_3
        - Merch / マーチ
  /api/shops/{shopId}/merchs/{merchId}:
    get:
      operationId: MerchsController_findOne
      summary: 指定のマーチを取得する
      description: 指定のマーチを1つ取得します。
      parameters:
        - *ref_2
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMerchDto'
        '400':
          description: ''
      tags: *ref_3
info:
  title: 物販(仮) Document
  description: 物販(仮) の API リファレンスです
  version: 0.0.1
  contact: {}
tags: []
servers: []
components:
  schemas:
    CreateShopDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          example: 結束バンド
      required:
        - name
    ShopModel:
      type: object
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: 結束バンド
      required:
        - id
        - name
    ResponseShopDto:
      type: object
      properties:
        shop:
          $ref: '#/components/schemas/ShopModel'
      required:
        - shop
    ResponseShopListDto:
      type: object
      properties:
        shops:
          type: array
          items:
            $ref: '#/components/schemas/ShopModel'
      required:
        - shops
    UpdateShopDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          example: 放課後ティータイム
    MerchModel:
      type: object
      properties:
        name:
          type: string
          example: Tシャツ2023
        description:
          type: string
          example: 2023年のTシャツ
        price:
          type: number
          example: 2500
        image:
          type: string
          example: https://static.bucket/path
        isPinning:
          type: boolean
          example: true
        shopId:
          type: number
          example: 1
        salesTaxRateId:
          type: number
          example: 1
      required:
        - name
        - description
        - price
        - image
        - isPinning
        - shopId
        - salesTaxRateId
    ResponsePaginationDto:
      type: object
      properties:
        cursor:
          type: number
          description: ページネーションのカーソル
          example: 34
      required:
        - cursor
    ResponseMerchListDto:
      type: object
      properties:
        merchs:
          type: array
          items:
            $ref: '#/components/schemas/MerchModel'
        pagination:
          $ref: '#/components/schemas/ResponsePaginationDto'
      required:
        - merchs
        - pagination
    ResponseMerchDto:
      type: object
      properties:
        merch:
          $ref: '#/components/schemas/MerchModel'
      required:
        - merch